Struct string_telephone::client::ClientConnectionConfig

pub struct ClientConnectionConfig {
    pub max_connect_retries: u32,
    pub connect_attempt_timeout: Duration,
}

Additional configuration options for a Client connection

Fields

max_connect_retries

How many times should we ask for a connection before giving up?

connect_attempt_timeout

How long should each connection request await an answer?

Methods

impl ClientConnectionConfig

fn new(max_connect_retries: u32, connect_attempt_timeout: Duration) -> ClientConnectionConfig

Create a new ClientConnectionConfig object